requirements.txt
anthropic==0.3.10
azure-ai-ml
hyperopt==0.2.7
joblib==1.2.0
langchain==0.0.271
matplotlib==3.2.1
numpy==1.21.6
openai==0.27.8
pandas==1.1.5
petname==2.6
pyspark==3.3.1
python-dotenv==1.0.0
scikit_learn==0.22.1
sentence_transformers==2.2.2
tenacity==8.2.2
torch==1.12.0
tqdm==4.65.0
transformers==4.16.0
pre-commit
detect-secrets
